backend/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
dnspython==2.8.0
gunicorn==25.3.0
idna==3.12
numpy==2.4.4
packaging==26.1
pyaml==26.2.1
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-whois==0.9.6
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
